Reduce timedelta for match being played
Reducing 2h delay to 1h15 after begining of match for it to be taken into account. Should be easier to do stats in between matches. Especially when a team plays two matches in a row. Might cause a bug if data is extracted right after 1h15 and replays from the match are not uploaded.